requirements.txt
discord.py==3.10.4